+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * New cop `FileName` makes sure that source files have snake_case names. ([@bbatsov][])
+* New cop `DeprecatedClassMethods` checks for deprecated class methods. ([@bbatsov][])
 * [#743](https://github.com/bbatsov/rubocop/issues/743): `SingleLineMethods` cop does auto-correction. ([@jonas054][])
 * [#743](https://github.com/bbatsov/rubocop/issues/743): `Semicolon` cop does auto-correction. ([@jonas054][])
 * [#743](https://github.com/bbatsov/rubocop/issues/743): `EmptyLineBetweenDefs` cop does auto-correction. ([@jonas054][])